code working, and everything hooked up, details to finish: make it responsive. revise data on graph,try to pass son file in to avoid clutter
